[Refactor] To use ProtectedRoute for custom pages
Custom pages rely a lot on the state data and order is important, so
there should be a reusable mechanism for checking if the state has been
loaded and if it contains matching information for the page.

Refactored. Now if there is no loaded state or no matching data,
redirects to home page.